refactor: CurrentTime 컴포넌트 분리 (#104)
