+++ src/components/PersonalInfo/PersonalInfo.js
@@ -1,5 +1,3 @@
-import dayjs from 'dayjs';
-
 import ProfileInfo from './ProfileInfo.js';
 import WorkInfo from './WorkInfo.js';
 import './PersonalInfo.css';
@@ -8,22 +6,6 @@ export default class PersonalInfo {
   constructor() {
     this.ProfileInfo = new ProfileInfo();
     this.WorkInfo = new WorkInfo();
-  }
-
-  updateTime() {
-    const currentTime = dayjs().format('HH:mm');
-
-    const $time = document.querySelector('.work-hour-time.current');
-    $time.setAttribute('datetime', currentTime);
-    $time.innerText = currentTime;
-  }
-
-  getNextUpdateDelay() {
-    const now = dayjs();
-    const nextMinuteStart = now.add(1, 'minute').startOf('minute');
-    const delay = nextMinuteStart.diff(now);
-
-    return delay;
   }
 
   render() {
